Removed utility and compat files, simplified stemmer and tokenizer, various reorganizations
