+++ etl/textrank_algorithm/nlp/__init__.py
@@ -1,2 +1,2 @@
 from .tokenizers import Tokenizer
-from .stemmers import *
+from .stemmer import *
