--- etl/textrank_algorithm/sumy/__init__.py
@@ -1,11 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from __future__ import absolute_import
-from __future__ import division, print_function, unicode_literals
-
-from .nlp import *
-from .summarizers import *
-from .utils import *
-
-__author__ = "Michal Belica"
-__version__ = "0.8.1"
